make cartUtils.js to handle all functions of localstorage for cart, make foodDetails page

+++ src/Pages/Dashboard/AddFood/AddFood.jsx
@@ -89,7 +89,7 @@ const AddFood = () => {
     const newFood = {
       name: data.name,
       price: parseFloat(data.price),
-      discount: parseInt(data.discount) || 0,
+      discount: parseFloat(data.discount) || 0,
       description: data.description,
       images: imageURLs,
       available: data.available === "true",
