+++ src/Pages/Dashboard/EditFood/EditFood.jsx
@@ -124,7 +124,7 @@ const EditFood = () => {
     const payload = {
       name: data.name,
       price: parseFloat(data.price),
-      discount: parseInt(data.discount) || 0,
+      discount: parseFloat(data.discount) || 0,
       description: data.description,
       available: data.available === "true",
       imagesToAdd: newImagesURLs,
